third_party/blink/web_tests/external/wpt/html/syntax/parsing/html5lib_tests7_run_type=uri-expected.txt

This is a testharness.js-based test.
[FAIL] html5lib_tests7.html 5b0b3edcc3ce9fdc9f58eb62d326865ca0aab8c8
  assert_equals: expected "#document\\n| <html>\\n|   <head>\\n|   <body>\\n|     <select>\\n|     <keygen>" but got "#document\\n| <html>\\n|   <head>\\n|   <body>\\n|     <select>\\n|       <keygen>"
Harness: the test ran to completion.
